+++ packages/demo/src/contacts/ContactListFilter.tsx
@@ -29,9 +29,10 @@ export const ContactListFilter = () => {
         sort: { field: 'name', order: 'ASC' },
     });
     return (
-        <Box width="13em" minWidth="13em" order={-1} mr={2} mt={7}>
+        <Box width="13em" minWidth="13em" order={-1} mr={2} mt={5}>
             <FilterLiveSearch
                 source="fts@fts"
+                hiddenLabel
                 sx={{
                     display: 'block',
                     '& .MuiFilledInput-root': { width: '100%' },