+++ packages/demo/src/notes/NewNote.tsx
@@ -10,7 +10,7 @@ import {
     Identifier,
     useResourceContext,
 } from 'react-admin';
-import { Box, TextField as TextInput, Button } from '@mui/material';
+import { Box, TextField as TextInput, Button, Stack } from '@mui/material';
 
 import { StatusSelector } from './StatusSelector';
 
@@ -40,8 +40,10 @@ export const NewNote = ({
             sales_id: identity.id,
             date,
             text,
-            type: 'Note',
         };
+        if (record.type) {
+            data.type = record.type;
+        }
         if (showStatus) {
             data.status = status;
         }
@@ -53,14 +55,13 @@ export const NewNote = ({
                     setText('');
                     notify('Note added successfully');
                     refetch();
-                    update(reference, {
-                        id: (record && record.id) as unknown as Identifier,
-                        data: {
-                            last_seen: date,
-                            status,
-                        },
-                        previousData: record,
-                    });
+                    if (reference === 'contacts') {
+                        update(reference, {
+                            id: (record && record.id) as unknown as Identifier,
+                            data: { last_seen: date, status },
+                            previousData: record,
+                        });
+                    }
                 },
             }
         );
@@ -84,7 +85,7 @@ export const NewNote = ({
                 <Box display="flex" justifyContent="space-between" mt={1}>
                     <span>
                         {text ? (
-                            <>
+                            <Stack direction="row">
                                 {showStatus && (
                                     <StatusSelector
                                         status={status}
@@ -102,6 +103,7 @@ export const NewNote = ({
                                     type="datetime-local"
                                     variant="filled"
                                     size="small"
+                                    margin="none"
                                     value={date}
                                     onChange={(
                                         event: React.ChangeEvent<HTMLInputElement>
@@ -115,7 +117,7 @@ export const NewNote = ({
                                         },
                                     }}
                                 />
-                            </>
+                            </Stack>
                         ) : null}
                     </span>
                     <Button