+++ packages/demo/src/notes/Note.tsx
@@ -8,7 +8,6 @@ import {
     useDelete,
     useUpdate,
     useNotify,
-    useRecordContext,
 } from 'react-admin';
 import {
     Box,
@@ -26,7 +25,6 @@ import { Status } from '../misc/Status';
 export const Note = ({
     showStatus,
     note,
-    isLast,
     reference,
 }: {
     showStatus?: boolean;
@@ -38,7 +36,6 @@ export const Note = ({
     const [isEditing, setEditing] = useState(false);
     const [noteText, setNoteText] = useState(note.text);
     const resource = useResourceContext();
-    const record = useRecordContext();
     const notify = useNotify();
 
     const [update, { isLoading }] = useUpdate();
@@ -50,11 +47,6 @@ export const Note = ({
             mutationMode: 'undoable',
             onSuccess: () => {
                 notify('Note deleted', { type: 'info', undoable: true });
-                update(reference, {
-                    id: record.id,
-                    data: { nb_notes: record.nb_notes - 1 },
-                    previousData: record,
-                });
             },
         }
     );